update: correcting portfolio stocks

+++ portfolio/apps/project/views.py
@@ -1,28 +1,28 @@
 import os
 from django.shortcuts import render, get_object_or_404, redirect
 from django.contrib import messages
+from django.db.models.signals import post_delete
+from django.dispatch import receiver
+
 from .models import Project
 from ..about.models import Tools
 from .forms import ProjectForm
-from django.db.models.signals import post_delete
-from django.dispatch import receiver
 
 
 def projects(request):
     filter_type = request.GET.get('type')
     if filter_type:
-        projects = Project.objects.filter(type=filter_type)
+        projects_qs = Project.objects.filter(type=filter_type)
     else:
-        projects = Project.objects.all()
+        projects_qs = Project.objects.all()
 
-    # Transformar tools em lista
-    for project in projects:
+    for project in projects_qs:
         project.tools_list = project.tools.split(',') if project.tools else []
 
     filters_list = ["Destaques", "RPA", "Dados", "Sistemas"]
 
     return render(request, 'projects/projects.html', {
-        'projects': projects,
+        'projects': projects_qs,
         'filters_list': filters_list,
         'filter_type': filter_type
     })
@@ -31,32 +31,44 @@ def projects(request):
 def read_projects(request):
     tools = Tools.objects.all().order_by('order')
     projects = Project.objects.all()
-    form = ProjectForm()
-    
-    return render(request, 'projects/create_projects.html', {
-        'form': form,
+
+    return render(request, 'projects/read_projects.html', {
         'tools': tools,
-        'projects': projects,
+        'projects': projects
+    })
+
+
+def view_project(request, pk):
+    current_project = get_object_or_404(Project, pk=pk)
+    tools_list = [tag.strip() for tag in current_project.tools.split(
+        ",")] if current_project.tools else []
+    projects_qs = Project.objects.filter(
+        type=current_project.type).exclude(pk=current_project.pk)
+
+    return render(request, 'projects/view_project.html', {
+        'current_project': current_project,
+        'projects': projects_qs,
+        'tools_list': tools_list
     })
 
 
 def create_project(request):
-    tools = Tools.objects.all().order_by('order')
-    form = ProjectForm(request.POST, request.FILES)
-    if form.is_valid():
-        project = form.save(commit=False)
-        selected_tools = request.POST.getlist("tools")
-        tools_str = ", ".join(selected_tools)
-        project.tools = tools_str
-        project.save()
-        messages.success(request, "Projeto adicionado com sucesso!")
-        return redirect('projects')
-    else:
-        for field, errors in form.errors.items():
-            for error in errors:
-                messages.error(request, f"{field}: {error}")
+    if request.method == 'POST':
+        form = ProjectForm(request.POST, request.FILES)
+        tools = Tools.objects.all().order_by('order')
+        projects = Project.objects.all()
 
-    return render(request, 'projects/create_projects.html', {
+        if form.is_valid():
+            project = form.save(commit=False)
+            selected_tools = request.POST.getlist("tools")
+            project.tools = ", ".join(selected_tools)
+            project.save()
+            messages.success(request, "Projeto adicionado com sucesso!")
+            return redirect('/projects/read/?status=success&msg=Projeto+adicionado+com+sucesso')
+        else:
+            return redirect(f'/projects/read/?status=error&msg=Erro+ao+atualizar+o+projeto\n{form.errors}')
+
+    return render(request, 'projects/read_projects.html', {
         'form': form,
         'tools': tools,
         'projects': projects,
@@ -72,54 +84,39 @@ def delete_project_files(sender, instance, **kwargs):
 
 
 def delete_project(request, pk):
-    project = get_object_or_404(Project, pk=pk)
-    project.delete()
-    return redirect('create_project')
-
-
-def view_project(request, pk):
-    current_project = get_object_or_404(Project, pk=pk)
-    tools_list = [tag.strip()
-                  for tag in current_project.tools.split(",")] if current_project.tools else []
-    projects = Project.objects.filter(type=current_project.type).exclude(pk=current_project.pk)
-    return render(request, 'projects/view_project.html', {
-        'current_project': current_project,
-        'projects': projects,
-        'tools_list': tools_list
-    })
+    if request.method == 'POST':
+        project = get_object_or_404(Project, pk=pk)
+        project.delete()
+        return redirect('/projects/read/?status=success&msg=Projeto+deletado+com+sucesso')
+    else:
+        return redirect('/projects/read/?status=error&msg=Erro+ao+deletar+o+projeto')
 
 
 def update_project(request, pk):
     project = get_object_or_404(Project, pk=pk)
+    tools = Tools.objects.all().order_by('order')
+    selected_tools = [tool.strip() for tool in project.tools.split(
+        ',')] if project.tools else []
 
     if request.method == 'POST':
-        # Cria um form apenas com os campos que podem ser editados
-        form = ProjectForm(
-            request.POST,
-            instance=project
-        )
-
+        form = ProjectForm(request.POST, request.FILES, instance=project)
         if form.is_valid():
-            # Atualiza apenas os campos permitidos
-            project.name = form.cleaned_data['name']
-            project.description = form.cleaned_data['description']
-            project.tools = ', '.join(request.POST.getlist('tools'))
-            project.link_project = form.cleaned_data['link_project']
-            project.type = form.cleaned_data['type']
-            project.save()
-
-            messages.success(request, "Projeto atualizado com sucesso!")
-            return redirect('projects')
+            updated_project = form.save(commit=False)
+            updated_project.name = form.cleaned_data['name']
+            updated_project.description = form.cleaned_data['description']
+            updated_project.type = form.cleaned_data['type']
+            updated_project.link_project = form.cleaned_data['link_project']
+            updated_project.tools = ', '.join(request.POST.getlist('tools'))
+            updated_project.save()
+            return redirect('/projects/read/?status=success&msg=Projeto+atualizado+com+sucesso')
+        else:
+            return redirect(f'/projects/read/?status=error&msg=Erro+ao+atualizar+o+projeto\n{form.errors}')
     else:
         form = ProjectForm(instance=project)
-        selected_tools = [tool.strip() for tool in project.tools.split(',')]
-
-    tools = Tools.objects.all().order_by('order')
 
     return render(request, 'projects/update_project.html', {
         'form': form,
         'tools': tools,
-        'editing': True,
-        'project': project,
         'selected_tools': selected_tools,
+        'project': project,
     })
